backend/requirements.txt
Django>=4.2,<5.0
django-cors-headers>=4.0,<5.0
djangorestframework>=3.15,<4.0
psycopg[binary]>=3.2,<4.0
requests>=2.32,<3.0
dj-database-url>=2.2,<3.0
